# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Outrun API",
    "description": "Wildfire evacuation dispatch for residents who can't evacuate alone. Every number comes from an engine; agents call these tools and explain results.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "paths": {
...
    "/clock/step": {
      "post": {
        "summary": "Advance the replay clock",
        "operationId": "step_clock",
        "parameters": [
          {
            "name": "n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 1,
              "title": "N"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClockInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/state": {
      "get": {
        "summary": "Everything the dashboard shows at time t",
        "operationId": "get_world_state",
        "parameters": [
          {
            "name": "t",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Replay time, ISO-8601 with offset, e.g. 2025-01-07T23:40:00-08:00. Omit to use the replay clock's current time.",
              "title": "T"
            },
            "description": "Replay time, ISO-8601 with offset, e.g. 2025-01-07T23:40:00-08:00. Omit to use the replay clock's current time."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/fire/outlook": {
      "get": {
        "summary": "Where the fire is likely to be in the next 1, 2 and 3 hours",
        "description": "Summary counts plus the most threatened map cells. Use it to describe the fire situation. All probabilities come from the nowcast engine.",
        "operationId": "get_fire_outlook",
        "parameters": [
          {
            "name": "t",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Replay time, ISO-8601 with offset, e.g. 2025-01-07T23:40:00-08:00. Omit to use the replay clock's current time.",
              "title": "T"
            },
            "description": "Replay time, ISO-8601 with offset, e.g. 2025-01-07T23:40:00-08:00. Omit to use the replay clock's current time."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutlookSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/residents/at-risk": {
      "get": {
        "summary": "Residents who need help, most urgent first",
        "description": "Each entry has the fire probability at their home, the pessimistic fire arrival, their rescue deadline and priority score, all computed by engines.",
        "operationId": "get_residents_at_risk",
        "parameters": [
          {
            "name": "t",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Replay time, ISO-8601 with offset, e.g. 2025-01-07T23:40:00-08:00. Omit to use the replay clock's current time.",
              "title": "T"
            },
            "description": "Replay time, ISO-8601 with offset, e.g. 2025-01-07T23:40:00-08:00. Omit to use the replay clock's current time."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AtRiskEntry"
                  },
                  "title": "Response Get Residents At Risk"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plan": {
      "get": {
        "summary": "The proposed rescue plan at time t",
        "description": "Routes per vehicle with ETAs, people no vehicle can reach in time (unreachable, must be escalated), and a reason per person.",
        "operationId": "get_current_plan",
        "parameters": [
          {
            "name": "t",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Replay time, ISO-8601 with offset, e.g. 2025-01-07T23:40:00-08:00. Omit to use the replay clock's current time.",
              "title": "T"
            },
            "description": "Replay time, ISO-8601 with offset, e.g. 2025-01-07T23:40:00-08:00. Omit to use the replay clock's current time."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/requests": {
      "post": {
        "summary": "Add a help request to the plan",
        "description": "Send a structured request: the caller's address, needs, people, notes. The backend looks up the coordinates and fills in the map cell, vehicle types and loading time. If the address isn't found, the error says what to ask the caller. The person becomes visible to the planner after the call-handling delay.",
        "operationId": "submit_help_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HelpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports": {
      "post": {
        "summary": "Add a report of fire at a place and time",
        "description": "Send a structured fire sighting: the nearest address, what was seen. The backend looks up the coordinates. It feeds the fire forecast from reported_at onward, alongside satellite detections.",
        "operationId": "submit_fire_report",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FireReportIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FireReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/decisions": {
      "get": {
        "summary": "All recorded decisions",
        "description": "Every approve, modify or reject decision recorded so far, in order.",
        "operationId": "list_decisions",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Decision"
                  },
                  "type": "array",
                  "title": "Response List Decisions"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Record what the emergency manager did with a plan",
        "description": "Call only after the emergency manager explicitly says approve, modify or reject, with the plan_id from get_current_plan. Never approve on your own.",
        "operationId": "record_decision",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Decision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Decision"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AtRiskEntry": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "needs": {
            "type": "string",
            "title": "Needs"
          },
          "people": {
            "type": "integer",
            "title": "People"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "p_1h": {
            "type": "number",
            "title": "P 1H"
          },
          "p_3h": {
            "type": "number",
            "title": "P 3H"
          },
          "arrival_p10_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrival P10 Min"
          },
          "deadline_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline Min"
          },
          "priority": {
            "type": "number",
            "title": "Priority"
          }
        },
        "type": "object",
        "required": [
          "id",
          "needs",
          "people",
          "lat",
          "lon",
          "source",
          "p_1h",
          "p_3h",
          "arrival_p10_min",
          "deadline_min",
          "priority"
        ],
        "title": "AtRiskEntry"
      },
      "ClockInfo": {
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          },
          "step_minutes": {
            "type": "integer",
            "title": "Step Minutes"
          },
          "now": {
            "type": "string",
            "format": "date-time",
            "title": "Now"
          }
        },
        "type": "object",
        "required": [
          "start",
          "end",
          "step_minutes",
          "now"
        ],
        "title": "ClockInfo"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OutlookSummary": {
        "properties": {
          "t": {
            "type": "string",
            "format": "date-time",
            "title": "T"
          },
          "wind_speed_ms": {
            "type": "number",
            "title": "Wind Speed Ms"
          },
          "wind_from_deg": {
            "type": "number",
            "title": "Wind From Deg"
          },
          "cells_burning": {
            "type": "integer",
            "title": "Cells Burning"
          },
          "cells_p1h_over_50": {
            "type": "integer",
            "title": "Cells P1H Over 50"
          },
          "cells_p3h_over_50": {
            "type": "integer",
            "title": "Cells P3H Over 50"
          },
          "top_cells": {
            "items": {
              "$ref": "#/components/schemas/FireOutlook"
            },
            "type": "array",
            "title": "Top Cells"
          }
        },
        "type": "object",
        "required": [
          "t",
          "wind_speed_ms",
          "wind_from_deg",
          "cells_burning",
          "cells_p1h_over_50",
          "cells_p3h_over_50",
          "top_cells"
        ],
        "title": "OutlookSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorldState": {
        "properties": {
          "t": {
            "type": "string",
            "format": "date-time",
            "title": "T"
          },
          "dataset": {
            "type": "string",
            "title": "Dataset"
          },
          "residents": {
            "items": {
              "$ref": "#/components/schemas/Resident"
            },
            "type": "array",
            "title": "Residents"
          },
          "vehicles": {
            "items": {
              "$ref": "#/components/schemas/Vehicle"
            },
            "type": "array",
            "title": "Vehicles"
          },
          "shelters": {
            "items": {
              "$ref": "#/components/schemas/Shelter"
            },
            "type": "array",
            "title": "Shelters"
          },
          "detections": {
            "items": {
              "$ref": "#/components/schemas/FireDetection"
            },
            "type": "array",
            "title": "Detections"
          },
          "reports": {
            "items": {
              "$ref": "#/components/schemas/FireReport"
            },
            "type": "array",
            "title": "Reports"
          },
          "facilities": {
            "items": {
              "$ref": "#/components/schemas/Facility"
            },
            "type": "array",
            "title": "Facilities"
          },
          "orders": {
            "items": {
              "$ref": "#/components/schemas/EvacOrder"
            },
            "type": "array",
            "title": "Orders"
          },
          "outlook": {
            "items": {
              "$ref": "#/components/schemas/FireOutlook"
            },
            "type": "array",
            "title": "Outlook"
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/ResidentRisk"
            },
            "type": "array",
            "title": "Risks"
          },
          "plan": {
            "$ref": "#/components/schemas/Plan"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/Decision"
            },
            "type": "array",
            "title": "Decisions"
          },
          "picked_up": {
            "additionalProperties": {
              "type": "string",
              "format": "date-time"
            },
            "type": "object",
            "title": "Picked Up",
            "description": "Resident id -> when a vehicle left with them"
          },
          "schedule": {
            "items": {
              "$ref": "#/components/schemas/Route"
            },
            "type": "array",
            "title": "Schedule",
            "description": "What each vehicle is actually doing: remaining approved stops, ETAs in minutes after t"
          },
          "plan_takes_effect_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Takes Effect At",
            "description": "When the plan above is approved and vehicles start following it"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "t",
          "dataset",
          "residents",
          "vehicles",
          "shelters",
          "detections",
          "reports",
          "facilities",
          "orders",
          "outlook",
          "risks",
          "plan",
          "decisions"
        ],
        "title": "WorldState",
        "description": "Everything the dashboard needs at time t, in one message."
      }
    }
  }
}